// sanity.types.ts
export type SanityImageCrop = {
  _type: "sanity.imageCrop";
  top?: number;
  bottom?: number;
  left?: number;
  right?: number;
};

export type SanityImageHotspot = {
  _type: "sanity.imageHotspot";
  x?: number;
  y?: number;
  height?: number;
  width?: number;
};
// ...
export declare const internalGroqTypeReferenceTo: unique symbol;
// Source: ./src/lib/sanityQueries.ts
// Variable: getAllProjects
// Query: *[_type == "projetos-blogPost"]| order(_createdAt asc)  {  ...,"urlforimage":projectImage.asset._ref, priority,  githubLink,  youtubeLink,  projectLink,  projectName,  "slug":slug.current,  tags[]->{    name  },  category[]->{    name  },  fullDescription}
export type GetAllProjectsResult = Array<{
  _id: string;
  _type: "projetos-blogPost";
  _createdAt: string;
  _updatedAt: string;
  _rev: string;
  projectName: string;
  priority: boolean | null;
  category: Array<{
    name: string;
  }> | null;
  tags: Array<{
    name: string;
  }> | null;
  projectImage: {
    asset?: {
      _ref: string;
      _type: "reference";
      _weak?: boolean;
      [internalGroqTypeReferenceTo]?: "sanity.imageAsset";
    };
    hotspot?: SanityImageHotspot;
    crop?: SanityImageCrop;
    alt?: string;
    _type: "image";
  };
  shortDescription: string;
  fullDescription: Array<{
    children?: Array<{
      marks?: Array<string>;
      text?: string;
      _type: "span";
      _key: string;
    }>;
    style?: "blockquote" | "h1" | "h2" | "h3" | "h4" | "h5" | "h6" | "normal";
    listItem?: "bullet" | "number";
    markDefs?: Array<{
      href?: string;
      _type: "link";
      _key: string;
    }>;
    level?: number;
    _type: "block";
    _key: string;
  }>;
  projectLink: string | null;
  githubLink: string | null;
  youtubeLink: string | null;
  slug: string;
  urlforimage: string | null;
}>;